+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 - The Request title can be entered in the toolbar now.
 - Improved save and delete request buttons. They use a dropdown menu for confirmation now, which is much faster to use.
 
+### Fixed
+- When using environments in a request, the request would show up on the history page with the currently active environment. Now it correctly shows up with the values used when executing the request.
+
 ## [1.14.0] - 2016-08-20
 ### Added
 - Splash screen.